Moved shortcuts to output, removed graphing.
