+++ modules/shortcuts.py
@@ -6,13 +6,38 @@
 
 
 import json
+import os
+import sys
+import typing as t
+from datetime import datetime
 from http.client import responses
-from typing import Union
 
+import numpy as np
 from aiohttp import ClientSession
 
 
-def gen_block(content: Union[str, list, dict], *,
+def path(*filepath) -> str:
+    """Returns absolute path from main caller file to another location.
+
+    Args:
+        filepath (iritable): Arguments to add to the current filepath.
+
+    Returns:
+        String of filepath with OS based seperator.
+
+    Examples:
+        >>> print(path('tmp', 'image.png'))
+        C:\\Users\\Xithr\\Documents\\Repositories\\Xythrion\\tmp\\image.png
+
+    """
+    lst = [
+        os.path.abspath(os.path.dirname(sys.argv[0])),
+        (os.sep).join(str(y) for y in filepath)
+    ]
+    return (os.sep).join(str(s) for s in lst)
+
+
+def gen_block(content: t.Union[str, list, dict], *,
               lang: str = 'py', lines: bool = False, separator: str = '|') -> str:
     """Generates a Discord markdown block.
 
@@ -58,7 +83,7 @@ def gen_block(content: Union[str, list, dict], *,
     return f'```{lang}\n{content}\n```'
 
 
-async def http_get(url: str, *, session: ClientSession = None, block: bool = False) -> Union[dict, str]:
+async def http_get(url: str, *, session: ClientSession = None, block: bool = False) -> t.Union[dict, str]:
     """Gets information from a http request.
 
     Args:
@@ -119,3 +144,96 @@ async def http_get(url: str, *, session: ClientSession = None, block: bool = Fal
             return gen_block(dict(info), lang='py' if content_type == 'json' else content_type, lines=True)
 
         return info
+
+
+def gen_filename() -> str:
+    """Generates a filename.
+
+    Returns:
+        A string with the current date for filename usage.
+
+    Examples:
+        >>> print(gen_filename())
+        1584774125328021
+
+        >>> print(f'{gen_filename()}.png')
+        1584774141733494.png
+
+    """
+    return str(datetime.timestamp(datetime.now())).replace('.', '')
+
+
+def get_extensions() -> list:
+    """Gets all the extensions within a folder for the bot to load.
+
+    Returns:
+        a list of cogs starting with cogs.<folder if any>.<filename without .py>.
+
+    Examples:
+        >>> print(get_extensions())
+        ['cogs.math.calculator', 'cogs.math.graphing', 'cogs.meta.custom']
+
+    """
+    c = []
+
+    for folder in os.listdir(path('cogs')):
+        c.extend([f'cogs.{folder}.{i[:-3]}' for i in os.listdir(path('cogs', folder)) if i[-3:] == '.py'])
+
+    return c
+
+
+def shorten(s: str, char_limit: int = 50) -> str:
+    """Cuts down a string to a specific length and adds '...'
+
+    Args:
+        s (str): The input string
+        char_limit (int): The desired limit for length
+
+    Returns:
+        A string cut down to a specific length.
+
+    Examples:
+        >>> print(shorten("A large string that's very very very long and pretty much never ending until now"))
+        A large string that's very very very long and pretty...
+
+        >>> print(shorten("A large string that's very very very long."))
+        A large string that's very very very long.
+
+    """
+    new_str = []
+    for word in s.strip().split():
+        if len(' '.join(str(y) for y in new_str)) < char_limit:
+            new_str.append(word)
+        else:
+            break
+
+    new_str = ' '.join(str(y) for y in new_str).strip()
+    return new_str + '...' if len(new_str) >= char_limit else new_str
+
+
+def gen_table(lst: t.List[list], *, vertical_sep: str = '|', horizontal_sep: str = '-',
+              title_data_gap: t.Union[str, bool] = '-',
+              boxed: bool = False) -> t.Union[str, list]:
+    """Generating a table for data.
+
+    Args:
+        lst (:obj:`typing.List[list]`): A 2D list containing table information.
+
+    Returns:
+
+    Examples:
+        >>> lst = [['column', 'column 2', 'column 3'], [1, 1, 1], [2, 2, 2], [3, 3, 3]]
+        >>> print(gen_table(lst))
+        column
+             1
+             2
+             3
+
+        >>> print()
+        column | column 2 | column 3
+        -------|----------|---------
+
+    """
+    lst = np.array(lst)
+    lst = [list(lst[:, i]) for i in range(np.shape(lst)[1])]
+    return [[y.rjust(max(map(len, x))) for y in x] for x in lst]
